requirements.txt
transformers==4.30.2
torch>=2.2.0
sentence-transformers==2.2.2
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pymupdf==1.22.3
chromadb==0.4.6
pydantic==1.10.8
python-dotenv==1.0.0
jinja2==3.1.2
numpy==1.24.3
google-generativeai>=0.3.0